Update fast to 1.0.1 and fix trying to delete not existing folder (#14)

+++ src/main/java/dev/lost/engine/assetsgenerators/DataPackGenerator.java
@@ -238,7 +238,8 @@ public class DataPackGenerator {
     }
 
     public void build(@NotNull File dataPackFolder) throws IOException {
-        FastFiles.deleteFolderSync(dataPackFolder.toPath());
+        if (dataPackFolder.exists() && dataPackFolder.isDirectory())
+            FastFiles.deleteFolderSync(dataPackFolder.toPath());
         saveJsonToFile(MCMETA,                   new File(dataPackFolder + "/pack.mcmeta"));
         saveJsonToFile(swords,                   new File(dataPackFolder + "/data/minecraft/tags/item/swords.json"));
         saveJsonToFile(pickaxes,                 new File(dataPackFolder + "/data/minecraft/tags/item/pickaxes.json"));
